backend/requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
python-dotenv
uvicorn
